Organized projects into Python, Java, and C folders
