+++ Java/Captcha_in_Java/src/main/resources/application.properties
