+++ Java/Captcha_in_Java/src/main/resources/static/login.js
