--- PYTHON PROJECTS/Typing-speed-test/README.md
@@ -1,32 +1,0 @@
-# Typing Speed Test
-
-A simple **Typing Speed Test** game implemented in Python.  
-Test your typing speed and accuracy with random sentences, and see your **Words Per Minute (WPM)**.
-
----
-
-## Features
-
-- Random text selection for typing practice
-- Measures typing time
-- Calculates typing accuracy
-- Computes Words Per Minute (WPM)
-- Console-based, lightweight, easy to run
-
----
-
-## Requirements
-
-- Python 3.x installed
-- No additional libraries required
-
----
-
-## How to Run
-
-1. Navigate to the project folder:
-
-```bash
-cd "Typing-speed-test"
-
-python Typing-speed-test.py
