+++ Python/Password-Manager/.gitignore
