+++ Python/Password-Manager/password_manager.py
