+++ Python/Password-Manager/requirements.txt
