+++ Python/life_game.py
