+++ Python/watermark_apply/app.py
