+++ Python/watermark_apply/output/watermarked_text.jpg
