+++ Python/watermark_apply/requirements.txt
