+++ Python/watermark_apply/utils/watermark.py
